Move parse to another file and a few nits.
